+++ R/check_covar_balance.R
@@ -4,14 +4,14 @@
 #' @description
 #' Checks the covariate balance of original population or pseudo population.
 #'
-#' @param pseudo_pop The generated pseudo population. In the following format:
-#'   - 1st column: outcome (Y)
-#'   - 2nd column: exposure (w)
-#'   - 3rd column: gps
-#'   - 4th column to the end: covariates (c)
+#' @param w A vector of observed continuous exposure variable.
+#' @param c A data.frame of observed covariates variable.
 #' @param ci_appr The causal inference approach.
-#' @param nthread The number of available threads.
 #' @param optimized_compile If TRUE, use optimized compile approach.
+#' @param counter_weight A weight vector in different situations. If the
+#' matching approach is selected, it is an integer data.table of counters.
+#' In the case of the weighting approach, it is weight data.table.
+#' @param nthread The number of available threads.
 #' @param ... Additional arguments passed to different models.
 #'
 #' @details
@@ -39,6 +39,8 @@
 #'mydata$region <- as.factor(region)
 #'mydata$cf5 <- as.factor(mydata$cf5)
 #'
+#'
+#'
 #'pseudo_pop <- generate_pseudo_pop(mydata$Y,
 #'                                  mydata$treat,
 #'                                  mydata[c("cf1","cf2","cf3","cf4","cf5","cf6","year","region")],
@@ -57,23 +59,39 @@
 #'                                  scale = 0.5,
 #'                                  nthread = 1)
 #'
-#'adjusted_corr_obj <- check_covar_balance(pseudo_pop$pseudo_pop,
+#'adjusted_corr_obj <- check_covar_balance(w = pseudo_pop$pseudo_pop[, c("w")],
+#'                                         c = pseudo_pop$pseudo_pop[ ,
+#'                                         pseudo_pop$covariate_cols_name,
+#'                                          with=FALSE],
+#'                                         counter = pseudo_pop$pseudo_pop[, c("counter_weight")],
 #'                                         ci_appr="matching",
 #'                                         nthread=1,
 #'                                         covar_bl_method = "absolute",
 #'                                         covar_bl_trs = 0.1,
 #'                                         covar_bl_trs_type = "mean",
-#'                                         optimized_compile=FALSE)
+#'                                         optimized_compile=TRUE)
 #'
 
-check_covar_balance <- function(pseudo_pop, ci_appr, nthread=1,
-                                optimized_compile, ...){
+check_covar_balance <- function(w,
+                                c,
+                                ci_appr,
+                                optimized_compile,
+                                counter_weight = NULL,
+                                nthread=1,
+                                  ...){
 
   # Passing packaging check() ----------------------------
   covar_bl_method <- NULL
   covar_bl_trs <- NULL
   covar_bl_trs_type <- NULL
   # ------------------------------------------------------
+
+  # logger::log_debug("Expecting w to be a data.table. ",
+  #                   "type of w is: {class(w)}")
+  # logger::log_debug("Expecting c to be a data.table. ",
+  #                   "type of c is: {class(c)}")
+  # logger::log_debug("Expecting counter_weight to be a data.table.",
+  #                   "type of counter_weight is: {class(counter_weight)}")
 
   logger::log_debug("Started checking covariate balance ... ")
   s_ccb_t <- proc.time()
@@ -86,33 +104,7 @@ check_covar_balance <- function(pseudo_pop, ci_appr, nthread=1,
     assign(i,unlist(dot_args[i], use.names = FALSE))
   }
 
-  if (ci_appr == 'adjust'){
-    # No covariate balance test for the 'adjust' causal inference approach.
-    stop("The code should never get here. Argument checks or while loop checks
-         are not correct.")
-  }
-
-  if (covar_bl_method == 'absolute'){
-    if (ci_appr == 'matching'){
-      if (!optimized_compile){
-        abs_cor <- absolute_corr_fun(pseudo_pop[, 2],
-                                     pseudo_pop[,6:length(pseudo_pop)])
-        #names(abs_cor$absolute_corr) <- names(pseudo_pop)[6:length(pseudo_pop)]
-      } else if (optimized_compile){
-        abs_cor <- absolute_weighted_corr_fun(pseudo_pop[, 2], pseudo_pop[, 4],
-                                     pseudo_pop[,6:length(pseudo_pop)])
-        #names(abs_cor$absolute_corr) <- names(pseudo_pop)[6:length(pseudo_pop)]
-      } else {
-        stop("The code should never get here. There is something wrong with check arguments.")
-      }
-    } else if (ci_appr == 'weighting') {
-      abs_cor <- absolute_weighted_corr_fun(pseudo_pop[, 2],pseudo_pop[, 6],
-                                            pseudo_pop[, 7:length(pseudo_pop)])
-      #names(abs_cor$absolute_corr) <- names(pseudo_pop)[7:length(pseudo_pop)]
-    } else {
-      stop(paste("Selected causal inference approach (ci_appr =", ci_appr,
-                 ") is not implemented."))
-    }
+  post_process_abs <- function(abs_cor){
 
     logger::log_debug(paste("Mean absolute correlation: ",
                             abs_cor$mean_absolute_corr))
@@ -132,8 +124,43 @@ check_covar_balance <- function(pseudo_pop, ci_appr, nthread=1,
     logger::log_debug("Finished checking covariate balance (Wall clock time:  ",
                       " {(e_ccb_t - s_ccb_t)[[3]]} seconds).")
     return(output)
-  } else {
-    stop(paste(covar_bl_method, " method for covariate balance is not a valid
-               option."))
+
   }
+
+  if (covar_bl_method != "absolute"){
+    stop(paste(covar_bl_method, " method for covariate balance is not a valid
+               option or not implemented."))
+  }
+
+  if (!(ci_appr %in% c("matching", "weighting"))){
+    stop(paste (ci_appr, " is not a valid causal inference approach."))
+  }
+
+  if (is.null(counter_weight)){
+    abs_cor <- absolute_corr_fun(w, c)
+    return(post_process_abs(abs_cor))
+  }
+
+  if (ci_appr == "matching"){
+    if (optimized_compile){
+      abs_cor <- absolute_weighted_corr_fun(w = w,
+                                            vw = counter_weight,
+                                            c = c)
+    } else {
+      abs_cor <- absolute_corr_fun(w, c)
+    }
+    return(post_process_abs(abs_cor))
+  }
+
+  if (ci_appr == "weighting"){
+    abs_cor <- absolute_weighted_corr_fun(w = w,
+                                          vw = counter_weight,
+                                          c = c)
+    return(post_process_abs(abs_cor))
+  }
+
+
+  stop(paste0("Input values for check_covar_balance are not correct.",
+       " The code should not get here. Please inform the developers."))
+
 }