+++ R/create_matching.R
@@ -120,7 +120,7 @@ create_matching <- function(dataset, bin_seq = NULL, gps_model = "parametric",
       added_count <- freq_table[["N"]]
       counter_tmp <- numeric(nrow(cp_original_data))
       counter_tmp[index_of_data] <- added_count
-      cp_original_data$counter <- counter_tmp
+      cp_original_data$counter_weight <- counter_tmp
     }
 
     e_comp_p <- proc.time()