+++ R/estimate_gps.R
@@ -130,9 +130,9 @@ estimate_gps <- function(Y,
 
   w_mx <- compute_min_max(w)
   gps_mx <- compute_min_max(gps)
-  counter <- (w*0)+0 # initialize counter.
+  counter_weight <- (w*0)+0 # initialize counter.
   row_index <- seq(1,length(w),1) # initialize row index.
-  dataset <- cbind(Y,w,gps,counter, row_index, c)
+  dataset <- cbind(Y,w,gps,counter_weight,row_index, c)
 
   # Logging for debugging purposes
   logger::log_debug("Min Max of treatment: {paste(w_mx, collapse = ', ')}")