+++ R/estimate_semipmetric_erf.R
@@ -45,26 +45,26 @@
 #'
 estimate_semipmetric_erf <- function(formula, family, data, ci_appr){
 
-  counter <- ipw <- NULL
+  counter_weight <- NULL
 
   if (ci_appr == "matching"){
 
     # If the approach is not optimized, the counter will be zero, which causes
     # problem in generating prediction model.
-    if (sum(data$counter) == 0) {
-      data$counter <- data$counter + 1
+    if (sum(data$counter_weight) == 0) {
+      data$counter_weight <- data$counter_weight + 1
       logger::log_debug("Giving equal weight for all samples.")
     }
 
     suppressWarnings(gam_model <- gam::gam(formula = formula,
                                            family = family,
                                            data = data,
-                                           weights = counter))
+                                           weights = counter_weight))
   } else if (ci_appr == "weighting"){
     suppressWarnings(gam_model <- gam::gam(formula = formula,
                                            family = family,
                                            data = data,
-                                           weights = ipw))
+                                           weights = counter_weight))
   } else {
     stop(paste("ci_appr: ", ci_appr, " is not a valid causal inference."))
   }